requirements.txt
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
colorama==0.4.6
ConfigParser==7.1.0
cryptography==43.0.3
# customtkinter==24.3.1
docutils==0.21.2
Faker==30.8.1
fastapi==0.115.4
filelock==3.16.1
flask_appbuilder==4.5.2
h2==4.1.0
HTMLParser==0.0.2
# ipython==8.29.0
ipywidgets==8.1.5
# jnius==1.1.0
joblib==1.4.2
js==1.0
keyring==25.5.0
matplotlib==3.9.2
numpy==2.1.2
pandas==2.2.3
Pillow==11.0.0
pyarrow==???
protobuf==5.28.3
# pyodide==0.0.2
pyOpenSSL==24.2.1
python-dotenv==1.0.1
redis==5.2.0
scikit_learn==1.5.2
seaborn==0.13.2
simplejson==3.19.3
Sphinx==8.1.3
thread==2.0.5
urllib3_secure_extra==0.1.0
# xmlrpclib==1.0.1
yellowbrick==1.5
zstandard==0.23.0
